Added fixes for Contao 4.9

+++ dca/tl_files.php
@@ -8,13 +8,17 @@
  * @license LGPL-3.0+
  */
 
-$disabledObjects = deserialize($GLOBALS['TL_CONFIG']['disabledTagObjects'], true);
+if (isset($GLOBALS['TL_CONFIG']['disabledTagObjects'])) {
+	$disabledObjects = deserialize($GLOBALS['TL_CONFIG']['disabledTagObjects'], true);
+} else {
+	$disabledObjects = array();
+}
 if (!in_array('tl_files', $disabledObjects))
 {
 	$GLOBALS['TL_DCA']['tl_files']['palettes']['default'] = str_replace(';meta', ';tags;meta', $GLOBALS['TL_DCA']['tl_files']['palettes']['default']);
 
 	$GLOBALS['TL_DCA']['tl_files']['config']['ondelete_callback'][] = array('tl_files_tags', 'removeContentElement');
-	$GLOBALS['TL_DCA']['tl_files']['config']['onload_callback'][] = array('tl_files_tags', 'onCopy');
+	$GLOBALS['TL_DCA']['tl_files']['config']['oncopy_callback'][] = array('tl_files_tags', 'onCopy');
 }
 
 $GLOBALS['TL_DCA']['tl_files']['fields']['tags'] = array
@@ -33,37 +37,32 @@ $GLOBALS['TL_DCA']['tl_files']['fields']['tags'] = array
  * @author     Helmut Schottmüller <https://github.com/hschottm>
  * @package    Controller
  */
-class tl_files_tags extends tl_files
+class tl_files_tags extends \Backend
 {
-	public function removeContentElement($dc)
-	{
-		$this->Database->prepare("DELETE FROM tl_tag WHERE from_table = ? AND tid = ?")
-			->execute($dc->table, $dc->id);
+	public function removeContentElement($source, \DataContainer $dc)
+	{ 
+        $fileModel = \FilesModel::findByPath($source);
+		if ($fileModel != null) {
+			$this->Database->prepare("DELETE FROM tl_tag WHERE from_table = ? AND tid = ?")
+			->execute($dc->table, $fileModel->id);
+		}
 	}
 
-	public function onCopy($dc)
-	{
-		if (is_array($this->Session->get('tl_files_copy')))
-		{
-			foreach ($this->Session->get('tl_files_copy') as $data)
-			{
-				$this->Database->prepare("INSERT INTO tl_tag (tid, tag, from_table) VALUES (?, ?, ?)")
-					->execute($dc->id, $data['tag'], $data['table']);
-			}
-		}
-		$this->Session->set('tl_files_copy', null);
-		if (\Input::get('act') != 'copy')
-		{
-			return;
-		}
-		$objTags = $this->Database->prepare("SELECT * FROM tl_tag WHERE tid = ? AND from_table = ?")
-			->execute(\Input::get('id'), $dc->table);
-		$tags = array();
-		while ($objTags->next())
-		{
-			array_push($tags, array("table" => $dc->table, "tag" => $objTags->tag));
-		}
-		$this->Session->set("tl_files_copy", $tags);
-	}
+    public function onCopy($source, $destination, \DataContainer $dc)
+    {
+        $sourceModel = \FilesModel::findByPath($source);
+        $destModel = \FilesModel::findByPath($destination);
+
+        if ($sourceModel != null && $destModel != null) {
+            $objTags = $this->Database->prepare("SELECT * FROM tl_tag WHERE tid = ? AND from_table = ?")->execute($sourceModel->id, $dc->table);
+            $tags = array();
+            while ($objTags->next()) {
+                \array_push($tags, array("table" => $dc->table, "tag" => $objTags->tag));
+            }
+            foreach ($tags as $entry) {
+                $this->Database->prepare("INSERT INTO tl_tag (tid, tag, from_table) VALUES (?, ?, ?)")->execute($destModel->id, $entry['tag'], $entry['table']);
+            }
+        }
+    }
 }
